E0/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
paho-mqtt==1.6.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6